+++ people-and-teams/profilecard/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/profilecard
 
+## 16.10.1
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 16.10.0
 
 ### Minor Changes